+++ assets/css/app.css
@@ -28,7 +28,7 @@ right:0;left:0;
 }
 .white-panel{
  background-color: rgba(255,255, 255, 1);
- height:500px;
+ height:700px;
  position:absolute;
  top:-50px;
  width:50%;
@@ -98,7 +98,7 @@ right:0;left:0;
  outline: none;
 }
 .login-show input[type="submit"] {
- max-width: 150px;
+ max-width: 250px;
  width: 100%;
  background: #444444;
  color: #f9f9f9;
@@ -123,7 +123,7 @@ right:0;left:0;
  outline: none;
 }
 .register-show input[type="submit"] {
- max-width: 150px;
+ max-width: 250px;
  width: 100%;
  background: #444444;
  color: #f9f9f9;
