+++ config/routes.yaml
@@ -1,3 +1,7 @@
+fos_user_accueil:
+    path: /
+    controller: App\Controller\SecurityController::loginAction
+
 fos_user:
     resource: "@FOSUserBundle/Resources/config/routing/all.xml"
 #    path: /
