+++ config/services.yaml
@@ -4,7 +4,7 @@
 # Put parameters here that don't need to change on each machine where the app is deployed
 # https://symfony.com/doc/current/best_practices/configuration.html#application-related-configuration
 parameters:
-    locale: 'en'
+    locale: 'fr'
 
 services:
     # default configuration for services in *this* file
@@ -26,3 +26,10 @@ services:
 
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
+    app.controller.register:
+          class: App\Controller\SecurityController
+          arguments:
+            $formFactory: '@fos_user.registration.form.factory'
+
+    FOS\UserBundle\Form\Factory\FactoryInterface:
+      alias: 'fos_user.registration.form.factory'
