--- src/UserBundle/User.php
@@ -1,14 +1,0 @@
-<?php
-// src/AppBundle/AppBundle.php
-
-namespace App\UserBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class UserBundle extends Bundle
-{
-    public function getParent()
-    {
-        return 'FOSUserBundle';
-    }
-}
