+++ templates/bundles/FOSUserBundle/ChangePassword/change_password_content.html.twig
@@ -1,8 +1,13 @@
 {% trans_default_domain 'FOSUserBundle' %}
 
+<div class="white-panel">
+  <div class="login-show">
+	<h2>Changer votre mot de passe</h2>
 {{ form_start(form, { 'action': path('fos_user_change_password'), 'attr': { 'class': 'fos_user_change_password' } }) }}
     {{ form_widget(form) }}
     <div>
         <input type="submit" value="{{ 'change_password.submit'|trans }}" />
     </div>
 {{ form_end(form) }}
+</div>
+</div>
