+++ templates/bundles/FOSUserBundle/Profile/edit_content.html.twig
@@ -1,8 +1,12 @@
 {% trans_default_domain 'FOSUserBundle' %}
-
-{{ form_start(form, { 'action': path('fos_user_profile_edit'), 'attr': { 'class': 'fos_user_profile_edit' } }) }}
+<div class="white-panel">
+  <div class="login-show">
+    <h2>Modifier votre compte</h2>
+    {{ form_start(form, { 'action': path('fos_user_profile_edit'), 'attr': { 'class': 'fos_user_profile_edit' } }) }}
     {{ form_widget(form) }}
     <div>
-        <input type="submit" value="{{ 'profile.edit.submit'|trans }}" />
+      <input type="submit" value="{{ 'profile.edit.submit'|trans }}"/>
     </div>
-{{ form_end(form) }}
+    {{ form_end(form) }}
+  </div>
+</div>
