+++ templates/bundles/FOSUserBundle/Profile/show_content.html.twig
@@ -1,6 +1,10 @@
 {% trans_default_domain 'FOSUserBundle' %}
-
-<div class="fos_user_user_show">
-    <p>{{ 'profile.show.username'|trans }}: {{ user.username }}</p>
-    <p>{{ 'profile.show.email'|trans }}: {{ user.email }}</p>
+<div class="white-panel">
+  <div class="login-show">
+    <h2>Votre compte</h2>
+    <div class="fos_user_user_show">
+      <p>{{ 'profile.show.username'|trans }}: {{ user.username }}</p>
+      <p>{{ 'profile.show.email'|trans }}: {{ user.email }}</p>
+    </div>
+  </div>
 </div>
