+++ templates/bundles/FOSUserBundle/Registration/confirmed.html.twig
@@ -1,0 +1,17 @@
+{% extends "@FOSUser/layout.html.twig" %}
+
+{% trans_default_domain 'FOSUserBundle' %}
+
+{% block fos_user_content %}
+  <div class="white-panel">
+    <div class="login-show">
+      	<h2>Validation de votre compte</h2>
+      <p>{{'registration.confirmed'|trans({'%username%': user.username}) }}</p>
+      {% if targetUrl %}
+        <p>
+          <a href="{{ targetUrl }}">{{'registration.back'|trans }}</a>
+        </p>
+      {% endif %}
+    </div>
+  </div>
+{% endblock fos_user_content %}
