+++ templates/bundles/FOSUserBundle/Registration/register_content.html.twig
@@ -1,8 +1,12 @@
 {% trans_default_domain 'FOSUserBundle' %}
-
-{{ form_start(form, {'method': 'post', 'action': path('fos_user_registration_register'), 'attr': {'class': 'fos_user_registration_register'}}) }}
+<div class="white-panel">
+  <div class="login-show">
+    <h2>Création de Compte</h2>
+    {{ form_start(form, {'method': 'post', 'action': path('fos_user_registration_register'), 'attr': {'class': 'fos_user_registration_register'}}) }}
     {{ form_widget(form) }}
     <div>
-        <input type="submit" value="{{ 'registration.submit'|trans }}" />
+      <input type="submit" value="{{ 'registration.submit'|trans }}"/>
     </div>
-{{ form_end(form) }}
+    {{ form_end(form) }}
+  </div>
+</div>
