+++ templates/bundles/FOSUserBundle/Security/login_content.html.twig
@@ -1,69 +1,43 @@
 {% trans_default_domain 'FOSUserBundle' %}
 
-
-
 <div class="login-info-box">
-	<h2>Have an account?</h2>
-	<p>Lorem ipsum dolor sit amet</p>
-	<label id="label-register" for="log-reg-show">Login</label>
-	<input type="radio" name="active-log-panel" id="log-reg-show"  checked="checked">
+	<h2>Connexion</h2>
+	<label id="label-register" for="log-reg-show">Connexion</label>
+	<input type="radio" name="active-log-panel" id="log-reg-show" checked="checked">
 </div>
 
 <div class="register-info-box">
-	<h2>Don't have an account?</h2>
-	<p>Lorem ipsum dolor sit amet</p>
-	<label id="label-login" for="log-login-show">Register</label>
+	<h2>Pas de compte?</h2>
+	<label id="label-login" for="log-login-show">S'enregistrer</label>
 	<input type="radio" name="active-log-panel" id="log-login-show">
 </div>
 
 <div class="white-panel">
 	<div class="login-show">
-		<h2>LOGIN</h2>
+
+		<h2>Connexion</h2>
 		{% if error %}
 		<div>{{ error.messageKey|trans(error.messageData, 'security') }}</div>
 		{% endif %}
-		<form action="{{ path("fos_user_security_check") }}" method="post">
+		<form action="{{ path("fos_user_security_check") }}"="fos_user_security_check" )=")" }}""="}}" "" method="post">
 			{% if csrf_token %}
-			<input type="hidden" name="_csrf_token" value="{{ csrf_token }}" />
+			<input type="hidden" name="_csrf_token" value="{{ csrf_token }}"/>
 			{% endif %}
-			<input type="text" id="username" name="_username" value="{{ last_username }}" required="required" autocomplete="username" />
-			<input type="password" id="password" name="_password" required="required" autocomplete="current-password" />
-			<input type="submit" id="_submit" name="_submit" value="{{ 'security.login.submit'|trans }}" />
+			<input type="text" id="username" name="_username" value="{{ last_username }}" required="required" autocomplete="username"/>
+			<input type="password" id="password" name="_password" required="required" autocomplete="current-password"/>
+			<input type="submit" id="_submit" name="_submit" value="{{ 'security.login.submit'|trans }}"/>
 		</form>
-		<a href="">Forgot password?</a>
+		<a href="">Mot de passe oublié?</a>
 	</div>
 	<div class="register-show">
-		<h2>REGISTER</h2>
+		<h2>Création de Compte</h2>
 		{% trans_default_domain 'FOSUserBundle' %}
 
 		{{ form_start(form, {'method': 'post', 'action': path('fos_user_registration_register'), 'attr': {'class': 'fos_user_registration_register'}}) }}
-		    {{ form_widget(form) }}
-		    <div>
-		        <input type="submit" value="{{ 'registration.submit'|trans }}" />
-		    </div>
+		{{ form_widget(form) }}
+		<div>
+			<input type="submit" value="{{ 'registration.submit'|trans }}"/>
+		</div>
 		{{ form_end(form) }}
-
-		<!-- <input type="text" placeholder="Email">
-		<input type="password" placeholder="Password">
-		<input type="password" placeholder="Confirm Password">
-		<input type="button" value="Register"> -->
 	</div>
 </div>
-
-
-<!-- <form action="{{ path("fos_user_security_check") }}" method="post">
-{% if csrf_token %}
-<input type="hidden" name="_csrf_token" value="{{ csrf_token }}" />
-{% endif %}
-
-<label for="username">{{ 'security.login.username'|trans }}</label>
-<input type="text" id="username" name="_username" value="{{ last_username }}" required="required" autocomplete="username" />
-
-<label for="password">{{ 'security.login.password'|trans }}</label>
-<input type="password" id="password" name="_password" required="required" autocomplete="current-password" />
-
-<input type="checkbox" id="remember_me" name="_remember_me" value="on" />
-<label for="remember_me">{{ 'security.login.remember_me'|trans }}</label>
-
-<input type="submit" id="_submit" name="_submit" value="{{ 'security.login.submit'|trans }}" />
-</form> -->
